Update CI to Handle no dev branch (#2)
To simplify the release workflow, the team has decided to remove the
`dev` branch. This change requires a few changes to the CI system for
all projects.

1. For every feature-branch pull request the is merged to `main`, create
an appropriately versioned RC release (a tag, a GitHub release, plus any
packaging) with the version indicated by the project's manifest file
followed by a `-N`, where `N` is an incrementing number based on the
number of RC tags for that release already in existence.
2. If the branch merging into `main` is a release branch, create an RC
for every push to that branch with the version indicated by the release
branch name followed by a `-N`, where `N` is an incrementing number
based on the number of RC tags for that release already in existence.
3. If a pull request is merged to `main` from a release branch, create a
full release with the version number indicated in the release branch
name.