Update Trebuchet Module Update Commands and Wait for Download Completion
We were using an old methodology for updating Trebuchet modules. This
update fixes that by using the built-in commands instead.

Additionally, in order to wait for the firmware package to be fully
downloaded to the instrument, use `clear_output_queue()` to write a
timestamp and read until we get it back.